rbd: record image-relative offset in object request
For an image object request we will need to know what offset within
the rbd image the request covers.  Record that when the object
request gets created.

Update the I/O error warnings so they use this so what's reported
is more informative.  Move the assignment of the offset and length
in rbd_request_fn() so they're always valid for the warning message.

Rename a local variable to fit the convention used everywhere else.

Signed-off-by: Alex Elder <elder@inktank.com>