rbd: update feature bits
There is a new rbd feature bit defined by the osd; add it to the
ones defined here.  Change RBD_FEATURES_ALL so it represents the
set of all feature bits rather than the ones we support.  Define
a new symbol RBD_FEATURES_SUPPORTED to indicate which ones we
support.

Signed-off-by: Alex Elder <elder@inktank.com>